test: add and move test
