+++ libs/ddre/src/lib/core/public_api.ts
@@ -1,2 +1,3 @@
 export type { DDREEngineInput } from './engine-input';
 export type { DDREEngineOutput } from './engine-output';
+export { InvalidInputError } from './errors/invalid-input.error';
